pyproject.toml
[tool.poetry]
name = "cologne-anmeldung-bot"
version = "0.1.0"
description = "A Telegram bot that checks for the earliest available schedule for Anmeldung in Cologne, Germany"
authors = ["Isagani Marquez <gani.marquez.jr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.4.2"
requests = "^2.28.0"
bs4 = "^0.0.1"
lxml = "^4.9.0"
python-telegram-bot = "^13.12"
loguru = "^0.6.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
